+++ src/components/link/link.stories.tsx
@@ -26,7 +26,8 @@ type Story = StoryObj<typeof Link>;
 export const Default: Story = {
     args: {
         children: 'Label',
-        href: '/test',
+        href: 'https://aragon.org',
+        target: '_blank',
     },
 };