+++ CHANGELOG.md
@@ -9,8 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `InputFileAvatar`, `EmptyState`, `Checkbox`, `CheckboxGroup`, `RadioGroup`, `Radio`, and `RadioCard`
-    components
+-   Implement `InputFileAvatar`, `CardEmptyState`, `EmptyState`, `Checkbox`, `CheckboxGroup`, `CheckboxCard`,
+    `RadioGroup`, `Radio`, and `RadioCard` components
+-   Export all component types
 
 ## [1.0.11] - 2024-02-06