+++ CHANGELOG.md
@@ -7,12 +7,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Implement `Tooltip` component
+
 ## [1.0.12] - 2024-02-12
 
 ### Added
 
--   Implement `CardEmptyState`, `EmptyState`, `Checkbox`, `CheckboxGroup`, `CheckboxCard`, `RadioGroup`, `Radio`,
-    `RadioCard`, and `Tooltip` components
+-   Implement `CardEmptyState`, `EmptyState`, `Checkbox`, `CheckboxGroup`, `CheckboxCard`, `RadioGroup`, `Radio`, and
+    `RadioCard` components
 -   Export all component types
 -   Handle style transitions on `ActionItem`, `Button` and `Toggle` components
 -   Property `defaultValue` to `ToggleGroup` component to set an initial value for uncontrolled usage